feat: implement 0x79 register data read/write command
- Add RegisterHandler to CommandHandlerRegistry for 0x79 command
- Implement ReadRegister and WriteRegister types with i16 data type
- Add read_register() and write_register() methods to HsesClient
- Update MockState to use i16 for register values
- Fix register data format: 2 bytes data + 2 bytes reserved = 4 bytes total
- Add register_operations example and integration tests
- Ensure compliance with HSES protocol specification

The register command supports:
- Register numbers: 0-999 (writable: 0-559)
- Data type: 16-bit signed integer (i16)
- Services: 0x0E (read), 0x10 (write)
- Payload: 4 bytes (2 bytes data + 2 bytes reserved)

+++ moto-hses-mock/src/state.rs
@@ -12,7 +12,7 @@ pub struct MockState {
     pub position: proto::Position,
     pub variables: HashMap<u8, Vec<u8>>,
     pub io_states: HashMap<u16, bool>,
-    pub registers: HashMap<u16, i32>,
+    pub registers: HashMap<u16, i16>,
     pub alarms: Vec<proto::Alarm>,
     pub alarm_history: AlarmHistory,
     pub current_job: Option<String>,
@@ -203,12 +203,12 @@ impl MockState {
     }
 
     /// Get register value
-    pub fn get_register(&self, reg_number: u16) -> i32 {
+    pub fn get_register(&self, reg_number: u16) -> i16 {
         self.registers.get(&reg_number).copied().unwrap_or(0)
     }
 
     /// Set register value
-    pub fn set_register(&mut self, reg_number: u16, value: i32) {
+    pub fn set_register(&mut self, reg_number: u16, value: i16) {
         self.registers.insert(reg_number, value);
     }
